+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-doc
 
+## 4.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.10.1
+
 ## 4.0.0
 
 ### Minor Changes