+++ packages/rocketh-signer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rocketh-signer
 
+## 4.0.1
+
+### Patch Changes
+
+- ask before proceeding option
+- Updated dependencies
+  - rocketh@0.10.1
+
 ## 4.0.0
 
 ### Patch Changes