+++ packages/rocketh-verifier/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-verifier
 
+## 0.9.1
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.10.1
+
 ## 0.9.0
 
 ### Minor Changes